fix: 修复getpeername的bug [AI Usage] 使用deepseek分析结果进行修改 LTP 测试 getpeername01 中存在两个导致 4 个测试用例失败的 bug：
  Bug 1：peer_addr() 在 socketpair 套接字上错误返回 ENOTCONN

  文件：os/src/net/unix.rs

  在 new_pair() 中，虽然 peer 连接字段被正确设置（表明套接字已连接），但 peer_addr 字段仍保持 Unnamed 状态。接着 peer_addr() 检查 peer_addr == Unnamed 并返回
  ENOTCONN，导致所有 socketpair 测试用例（EINVAL、EFAULT）在地址/长度校验之前就提前失败了。

  修复：将检查条件改为基于 self.inner.peer 连接状态，而非 peer_addr 值。即使 peer_addr 为 Unnamed（适用于未绑定的 socketpair），只要套接字已连接，getpeername
  也应能成功。

  Bug 2：addrlen 指针未被解引用

  文件：os/src/syscall/net/name.rs、os/src/syscall/mod.rs

  内核将 args[2]（即 socklen_t* 指针的地址）当作长度值使用——这是错误的。对于 getpeername(fd, addr, addrlen)，第三个参数是一个指向 socklen_t
  的指针，而内核错误地使用了指针地址，并未读取其指向的实际长度值。

  内核将 args[2]（即 socklen_t* 指针的地址）当作长度值使用——这是错误的。对于 getpeername(fd, addr, addrlen)，第三个参数是一个指向 socklen_t
  的指针，而内核错误地使用了指针地址，并未读取其指向的实际长度值。

  修复：
  - 新增 read_addrlen_from_user() 函数，用于从用户空间解引用 socklen_t* 指针
  - 若指针无效（NULL 或未映射）则返回 EFAULT
  - 若值以 i32 格式读取为负数（Linux 行为）则返回 EINVAL
  - sys_getsockname 和 sys_getpeername 现均以 usize（原始指针）方式接收 addrlen，并在内部进行解引用

+++ os/src/syscall/mod.rs
@@ -533,9 +533,9 @@ pub fn syscall(syscall_id: usize, args: [usize; 6]) -> SyscallRet {
         #[cfg(feature = "net")]
         Syscall::Connect => sys_connect(args[0], args[1] as *const u8, args[2] as u32),
         #[cfg(feature = "net")]
-        Syscall::GetSockName => sys_getsockname(args[0], args[1] as *const u8, args[2] as usize),
+        Syscall::GetSockName => sys_getsockname(args[0], args[1] as *const u8, args[2]),
         #[cfg(feature = "net")]
-        Syscall::GetPeerName => sys_getpeername(args[0], args[1] as *const u8, args[2] as u32),
+        Syscall::GetPeerName => sys_getpeername(args[0], args[1] as *const u8, args[2]),
         #[cfg(feature = "net")]
         Syscall::SendTo => sys_sendto(
             args[0],
